app/src/com/constellio/app/modules/rm/model/calculators/FolderCopyRulesExpectedTransferDatesCalculator.java
/*Constellio Enterprise Information Management

Copyright (c) 2015 "Constellio inc."

This program is free software: you can redistribute it and/or modify
it under the terms of the GNU Affero General Public License as
published by the Free Software Foundation, either version 3 of the
License, or (at your option) any later version.

This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU Affero General Public License for more details.

You should have received a copy of the GNU Affero General Public License
along with this program. If not, see <http://www.gnu.org/licenses/>.
*/
package com.constellio.app.modules.rm.model.calculators;

import java.util.Arrays;
import java.util.List;

import org.joda.time.LocalDate;

import com.constellio.app.modules.rm.RMConfigs;
import com.constellio.app.modules.rm.model.CopyRetentionRule;
import com.constellio.app.modules.rm.wrappers.Folder;
import com.constellio.model.entities.calculators.CalculatorParameters;
import com.constellio.model.entities.calculators.MetadataValueCalculator;
import com.constellio.model.entities.calculators.dependencies.ConfigDependency;
import com.constellio.model.entities.calculators.dependencies.Dependency;
import com.constellio.model.entities.calculators.dependencies.LocalDependency;

public class FolderCopyRulesExpectedTransferDatesCalculator
		extends AbstractFolderCopyRulesExpectedDatesCalculator
		implements MetadataValueCalculator<List<LocalDate>> {

	LocalDependency<LocalDate> decommissioningDateParam = LocalDependency.toADate(Folder.DECOMMISSIONING_DATE);
	LocalDependency<LocalDate> actualTransferDateParam = LocalDependency.toADate(Folder.ACTUAL_TRANSFER_DATE);

	ConfigDependency<Integer> configNumberOfYearWhenVariableDelayPeriod =
			RMConfigs.CALCULATED_SEMIACTIVE_DATE_NUMBER_OF_YEAR_WHEN_VARIABLEPERIOD.dependency();

	@Override
	protected List<? extends Dependency> getCopyRuleDateCalculationDependencies() {
		return Arrays.asList(decommissioningDateParam, actualTransferDateParam,
				configNumberOfYearWhenVariableDelayPeriod);
	}

	@Override
	protected LocalDate calculateForCopyRule(int index, CopyRetentionRule copyRule, CalculatorParameters parameters) {

		LocalDate decommissioningDate = parameters.get(decommissioningDateParam);
		LocalDate actualTransferDate = parameters.get(actualTransferDateParam);

		int numberOfYearWhenVariableDelay = parameters.get(configNumberOfYearWhenVariableDelayPeriod);

		if (decommissioningDate == null || actualTransferDate != null) {
			return null;
		}

		if (copyRule.getActiveRetentionPeriod().isVariablePeriod()) {
			if (numberOfYearWhenVariableDelay == -1) {
				return null;
			} else {
				return decommissioningDate.plusYears(numberOfYearWhenVariableDelay);
			}
		} else {
			return decommissioningDate.plusYears(copyRule.getActiveRetentionPeriod().getFixedPeriod());
		}

	}
}